feat(orchestrator.sequential)!: redefine api and behaviour
In the interface: we add a predict method. Also, backwards return Self, not
Layermap, this allows the user to deal with the orchestrator or layermap
transparently/interchangeably. Also, the step method does not affect the last
row of the layermap, due to the new addition of predict.

BREAKING_CHANGE: return type in backward has changed.