feat: replace forward_only arg with subset. allowed target_state
now instead of forward_only in layer_map enumeration we can have "all", "forward"
and "backward". In the same spirit we deprecated the step_inference function
and replaced it with filter_messages arg in step.
also, backward method now supports a different target state (the y in the
perceptron rule)